+++ LookupAnything/docs/release-notes.md
@@ -7,6 +7,7 @@
 * Fixed error clicking 'see also' crop link from a seed lookup.
 * Fixed support for custom locations with farm animals.
 * Fixed 'tile lookup' option in Generic Mod Config Menu UI not working correctly.
+* Fixed lookup error if another mod added invalid recipe data.
 * Fixed transparency issue with faded icons in the lookup menu (thanks to TehPers!).
 * Improved translations. Thanks to ChulkyBow (added Ukrainian), Evexyron (updated Spanish), mezen (updated German), and Zangorr (updated Polish)!